Refactor test files to match project's standards
