+++ test/runtests.jl
@@ -9,6 +9,7 @@ using FixedPointNumbers
 using Colors
 using Dates
 using Unitful
+import JSON3
 import ReadVTK
 import GeoInterface as GI
 import Shapefile as SHP
@@ -23,19 +24,10 @@ islinux = Sys.islinux()
 datadir = joinpath(@__DIR__, "data")
 savedir = mktempdir()
 
-# Note: Shapefile.jl saves Chains and Polygons as Multi
-# This function is used to work around this problem
-_isequal(d1::Domain, d2::Domain) = all(_isequal(g1, g2) for (g1, g2) in zip(d1, d2))
-
-_isequal(g1, g2) = g1 == g2
-_isequal(m1::Multi, m2::Multi) = m1 == m2
-_isequal(g, m::Multi) = _isequal(m, g)
-function _isequal(m::Multi, g)
-  gs = parent(m)
-  length(gs) == 1 && first(gs) == g
-end
+include("testutils.jl")
 
 testfiles = [
+  "utils.jl",
   "io/images.jl",
   "io/stl.jl",
   "io/obj.jl",
@@ -56,8 +48,7 @@ testfiles = [
   "formats.jl",
   "convert.jl",
   "gis.jl",
-  "noattrs.jl",
-  "utils.jl"
+  "noattrs.jl"
 ]
 
 @testset "GeoIO.jl" begin
